requirements.txt
Flask==2.0.1
onnxruntime==1.9.0
Pillow==8.3.2
numpy==1.21.4

